Problems with the use of the compare function
The compare function returns -1 if this.password is less than
hashPassword(arguments.password, this.salt), 0 if they are equal and 1
if the first case is reversed. this url shows that

http://help.adobe.com/en_US/ColdFusion/9.0/Developing/WSc3ff6d0ea77859461172e0811cbec09af4-7fd0.html

In Boolean expressions, True, nonzero numbers, and the strings “Yes”,
“1”, “True” are equivalent; and False, 0, and the strings “No”, “0”, and
“False” are equivalent.

Based on this, using the compare won't be the best thing to do.